airport

+++ src/views/airport/index.vue
@@ -1,51 +1,48 @@
 <template>
   <div class="app-container">
-    <el-container>
-      <el-header>
-        <el-button type="text" @click="dialogVisible = true">添加</el-button>
-      </el-header>
-      <el-main>
-        <el-table :data="tableData" style="width: 100%">
-          <el-table-column
-            prop="code"
-            label="三字码"
-            width="180"
-          ></el-table-column>
-          <el-table-column
-            prop="name"
-            label="机场名称"
-            width="300"
-          ></el-table-column>
-          <el-table-column
-            prop="city"
-            label="机场所在城市"
-            width="300"
-          ></el-table-column>
-          <el-table-column
-            label="操作"
-            width="100">
-            <template slot-scope="scope">
-              <el-button @click="edit(scope.row)" type="text" size="small">编辑</el-button>
-              <el-button @click="removeOne(scope.row.id)" type="text" size="small">删除</el-button>
-            </template>
-          </el-table-column>
-        </el-table>
-        <h-page-footer>
-          <el-pagination
-            @size-change="handleSizeChange"
-            @prev-click="prevClick"
-            @next-click="nextClick"
-            :current-page="currentPage"
-            background
-            layout="total,sizes,prev,next"
-            prev-text="上一页"
-            next-text="下一页"
-            :page-size="pageSize"
-            :total="total">
-          </el-pagination>
-        </h-page-footer>
-      </el-main>
-    </el-container>
+    <el-header>
+
+    </el-header>
+    <el-form :inline="true" :model="searchForm">
+      <el-button type="primary" @click="dialogVisible = true">添加</el-button>
+    </el-form>
+    <el-table :data="tableData" style="width: 100%">
+      <el-table-column
+        prop="code"
+        label="三字码"
+        width="180"
+      ></el-table-column>
+      <el-table-column
+        prop="name"
+        label="机场名称"
+        width="300"
+      ></el-table-column>
+      <el-table-column
+        prop="city"
+        label="机场所在城市"
+        width="300"
+      ></el-table-column>
+      <el-table-column
+        label="操作"
+        width="200">
+        <template slot-scope="scope">
+          <el-button @click="edit(scope.row)" type="primary" size="small">编辑</el-button>
+          <el-button @click="removeOne(scope.row.id)" type="danger" size="small">删除</el-button>
+        </template>
+      </el-table-column>
+    </el-table>
+      <el-pagination
+        @size-change="handleSizeChange"
+        @prev-click="prevClick"
+        @next-click="nextClick"
+        :current-page="currentPage"
+        background
+        layout="total,sizes,prev,next"
+        prev-text="上一页"
+        next-text="下一页"
+        :page-size="pageSize"
+        :total="total">
+      </el-pagination>
     <el-dialog title="机场信息" :visible.sync="dialogVisible" width="30%">
       <el-form ref="form" :model="form" label-width="110px">
         <el-form-item label="三字码">
@@ -71,6 +68,7 @@
     name: 'airport',
     data() {
       return {
+        searchForm:{},
         form: {
           code: '',
           name: '',
@@ -81,14 +79,18 @@
         lastId: '0',
         pageFlag: 'next',
         pageSize: 10,
-        total:0,
-        currentPage:0
+        total: 0,
+        currentPage: 0
       };
     },
     methods: {
-      loadData(lastId, pageSize,pageFlag) {
+      add:function(){
+        this.dialogVisible= true;
+        this.$refs['form'].resetFields();
+      },
+      loadData(lastId, pageSize, pageFlag) {
         this.$store
-          .dispatch('airport/list', {pageSize, lastId,pageFlag})
+          .dispatch('airport/list', {pageSize, lastId, pageFlag})
           .then(data => {
             this.tableData = data;
           })
@@ -96,7 +98,7 @@
             console.log(error);
           });
       },
-      loadTotal:function(){
+      loadTotal: function () {
         this.$store
           .dispatch('airport/total', {})
           .then(data => {
@@ -108,36 +110,36 @@
       },
       handleSizeChange: function (pageSize) {
         this.pageSize = pageSize;
-        this.loadData('0', this.pageSize,this.pageFlag);
+        this.loadData('0', this.pageSize, this.pageFlag);
       },
-      prevClick:function(){
+      prevClick: function () {
         this.pageFlag = 'prev';
         this.lastId = this.tableData[0].id;
         this.loadData(this.lastId, this.pageSize);
       },
-      nextClick:function(){
+      nextClick: function () {
         this.pageFlag = 'next';
         this.lastId = this.tableData[this.tableData.length - 1].id;
-        this.loadData(this.lastId, this.pageSize,this.pageFlag);
+        this.loadData(this.lastId, this.pageSize, this.pageFlag);
       },
       removeOne: function (id) {
-        // this.$confirm("此操作将永久删除该文件, 是否继续?", "提示", {
-        //   confirmButtonText: "确定",
-        //   cancelButtonText: "取消",
-        //   type: "warning"
-        // }).then(() => {
-        //   console.log("确定！"+id);
-        // }).catch(() => {
-        //   console.log("取消！"+id);
-        // });
-        this.$store
-          .dispatch('airport/removeOne', id)
-          .then(data => {
-            console.log(data);
-          })
-          .catch(error => {
-            console.log(error);
-          });
+        this.$confirm('是否确定删除机场信息?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          this.$store
+            .dispatch('airport/removeOne', id)
+            .then(data => {
+              console.log(data);
+              this.loadData('0', this.pageSize,this.pageFlag);
+            })
+            .catch(error => {
+              console.log(error);
+            });
+        }).catch(err => {
+          console.error(err)
+        })
       },
       edit: function (row) {
         this.dialogVisible = true;
@@ -151,6 +153,7 @@
           .dispatch('airport/save', this.form)
           .then(data => {
             console.log(data);
+            this.loadData('0', this.pageSize,this.pageFlag);
           })
           .catch(error => {
             console.log(error);
@@ -160,7 +163,7 @@
     },
     mounted() {
       console.log("load:cure:" + this.lastId + ",pageSize+" + this.pageSize);
-      this.loadData(this.lastId, this.pageSize);
+      this.loadData(this.lastId, this.pageSize,this.pageFlag);
       this.loadTotal();
     }
   };
